requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
# Для Selenium (если потребуется позже)
# selenium>=4.15.0
# webdriver-manager>=4.0.1
# Для загрузки в БД (например, PostgreSQL)
psycopg2-binary>=2.9.7